@@ -26,12 +26,11 @@ apply plugin: 'com.android.library'
 //}
 
 android {
-  compileSdkVersion 23
-  buildToolsVersion "23.0.3"
+  compileSdkVersion 28
 
   defaultConfig {
     minSdkVersion 16
-    targetSdkVersion 23
+    targetSdkVersion 28
     versionCode 5
     versionName "1.5"
   }
@@ -47,13 +46,13 @@ dependencies {
   compile 'com.thoughtbot:expandablerecyclerview:1.3'
 
   //android
-  compile 'com.android.support:appcompat-v7:23.4.0'
-  compile 'com.android.support:recyclerview-v7:23.4.0'
+  implementation 'androidx.appcompat:appcompat:1.1.0'
+  implementation 'androidx.recyclerview:recyclerview:1.1.0'
 
   //unit tests
-  testCompile 'junit:junit:4.12'
-  testCompile 'org.robolectric:robolectric:3.2.1'
-  testCompile 'org.mockito:mockito-core:1.10.5'
+  testImplementation 'junit:junit:4.12'
+  testImplementation 'org.robolectric:robolectric:3.2.1'
+  testImplementation 'org.mockito:mockito-core:1.10.5'
 }
 
 //apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
