@@ -1,6 +1,6 @@
 package com.thoughtbot.expandablerecyclerview.listeners;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import com.thoughtbot.expandablerecyclerview.viewholders.GroupViewHolder;
 
 public interface OnGroupClickListener {
