@@ -1,6 +1,6 @@
 package com.thoughtbot.expandablerecyclerview.viewholders;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import com.thoughtbot.expandablerecyclerview.models.ExpandableGroup;
 
