@@ -1,10 +1,12 @@
 package com.thoughtbot.expandablerecyclerview.sample.multitype;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.thoughtbot.expandablerecyclerview.sample.R;
 
 import static com.thoughtbot.expandablerecyclerview.sample.GenreDataFactory.makeGenres;
